Fix JSON string comparison.
Fix when a JSON is compared with a string, and apply
unquote as MySQL does.
